Refactor workflow to separate module
