+++ src/databricks/labs/ucx/assessment/workflows.py
@@ -1,5 +1,6 @@
 from databricks.labs.ucx.contexts.workflow_task import RuntimeContext
-from databricks.labs.ucx.framework.tasks import Workflow, job_task
+from databricks.labs.ucx.framework.tasks import job_task
+from databricks.labs.ucx.framework.workflow import Workflow
 
 
 class Assessment(Workflow):
