+++ src/databricks/labs/ucx/runtime.py
@@ -7,7 +7,8 @@ from pathlib import Path
 from databricks.labs.ucx.__about__ import __version__
 from databricks.labs.ucx.assessment.workflows import Assessment, DestroySchema
 from databricks.labs.ucx.contexts.workflow_task import RuntimeContext
-from databricks.labs.ucx.framework.tasks import Task, TaskLogger, Workflow, parse_args
+from databricks.labs.ucx.framework.tasks import Task, TaskLogger, parse_args
+from databricks.labs.ucx.framework.workflow import Workflow
 from databricks.labs.ucx.hive_metastore.workflows import (
     MigrateTablesInMounts,
     TableMigration,
